Use rtrb queue to schedule UI events
This means events are pushed into the queue outside the audio thread, so the
audio thread stays realtime-safe. This seems to have fixed the clicks I was
hearing.

An exception to this is the get_* events, which have to return a value to the
UI. But, those aren't called very often so I'm not worrying about them for now.